Massive reorganization of the files into folder structure
We put all the QMD files under their own directory structure. The directory contains the QMD file and also an images directory, so each chapter is now self-contained

As part of this reorg, I updated the links to images etc. to make sure they are all pointing to the right location