Distributed references - each chapter has its own references files
The references.bib file causes significant conflicts when multiple people work on their chapters. Therefore, we are dividing the references.bib file into individual chapters. This approach has its own drawbacks, such as the potential duplication of references. However, we can implement checks for this in the future. For now, this method reduces merge conflicts.