Summary
Additionally, the spacing between the heading number and the text has been corrected. The spacing was set to 10pt (which is too large and looks unattractive). This spacing is usually set to 0.5em, which is what I have done.

I have edited the chapters up to and including dl_primer.qmd.

Two issues that were assigned to me on GitHub have been fixed: Missing tag (```) closure #608."