Introduces privacy and security concepts
Explains the importance of security and privacy in machine learning systems, highlighting their roles in protecting data, models, and infrastructure.

Defines security as protecting against unauthorized access and privacy as protecting sensitive information.

Discusses the interdependent and trade-off-laden nature of security and privacy and emphasizes the need for comprehensive and adaptive defenses.