Improves LaTeX command robustness.
Protects LaTeX commands by using `\protect` to ensure proper expansion and prevent errors, especially when used in environments where arguments might be prematurely evaluated.

Specifically, it ensures correct handling of potentially fragile arguments in `\fbxIncludeIcon` and the `title` argument of the `tcolorbox` environment. Additionally, it corrects the command which was incorrectly replacing all hyphens instead of just the first one.